course controller fatto, manca course rating service impl e controller

+++ src/main/java/it/unikey/databaseassessmentrefactoring/DAL/entity/CourseEntity.java
@@ -20,7 +20,7 @@ public class CourseEntity {
 
     @Column(nullable = false)
     @NonNull
-    private Integer duration ;
+    private Double duration ;
 
     @OneToMany(mappedBy = "course")
     Set<CourseRatingEntity> ratings;
